src/dashboard/requirements.txt
streamlit
google-cloud-firestore
pandas
google-auth>=2.0.0